backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "matplotlib>=3.4.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pillow>=8.3.0",
    "python-multipart>=0.0.26",
    "scikit-learn>=0.24.0",
    "seaborn>=0.11.0",
    "tensorflow>=2.10.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.46.0",
]
